+++ CHANGELOG.md
@@ -8,6 +8,8 @@
  - Added a [new example](https://github.com/lovasoa/SQLpage/tree/main/examples/CRUD%20-%20Authentication) to the documentation
  - Bug fix: points with a latitude of 0 are now displayed correctly on the map component. 
  - Bug fix: in sqlite, lower(NULL) now returns NULL instead of an empty string. This is consistent with the standard behavior of lower() in other databases. SQLPage has its own implementation of lower() that supports unicode characters, and our implementation now matches the standard behavior of lower() in mainstream SQLite. 
+ - Allow passing data from the database to sqlpage functions. This fixes most errors like: `Arbitrary SQL expressions as function arguments are not supported.`.
+ - Better error messages in the dynamic component when properties are missing.
 
 ## 0.24.0 (2024-06-23)
  - in the form component, searchable `select` fields now support more than 50 options. They used to display only the first 50 options.